requirements.txt
Django==1.6.2
Markdown==2.4
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
dj-database-url==0.2.2
django-debug-toolbar==1.0.1
djorm-ext-pgarray==0.10
djrill==0.9.0
gunicorn==18.0
psycopg2==2.5.2
recaptcha-client==1.0.6
requests==2.2.1
sqlparse==0.1.11
tornado==3.2
wsgiref==0.1.2
